created folders for the remaining components and moved files in appropriate sections
